requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.2.1
fastapi==0.85.1
h11==0.14.0
idna==3.4
pydantic==1.10.2
pymongo==4.3.2
python-decouple==3.6
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
